changed floats to doubles

+++ Main.cpp
@@ -4,6 +4,6 @@ int main()
 {
     std::vector<int> Nlist{4};
     int Iterations = 30;
-    float Beta = 2.3;
+    double Beta = 2.3;
     Graph1(Nlist, Iterations, Beta);
 }
