+++ Util.cpp
@@ -1,18 +1,18 @@
 #include "Util.h"
 std::random_device rd;
 std::default_random_engine generator(rd());
-std::uniform_real_distribution<float> unidistribution(0, 1);
+std::uniform_real_distribution<double> unidistribution(0, 1);
 
-float Random()
+double Random()
 {
-    float R;
+    double R;
     for (int i = 0; i < 10; ++i)
     {
         R = unidistribution(generator);
     }
     return R;
 }
-int Choice(float proba1)
+int Choice(double proba1)
 {
     std::discrete_distribution<int> discrdistribution{proba1, (1 - proba1)};
     // auto gen = std::bind(std::discrete_distribution<int>{1, 0.5}, std::default_random_engine(rd()));
@@ -24,31 +24,31 @@ int Choice(float proba1)
     return b;
 }
 
-std::vector<float> get_ini_rand_elem()
+std::vector<double> get_ini_rand_elem()
 {
-    std::vector<float> elem;
+    std::vector<double> elem;
     elem.push_back(Random() * 2 - 1);
     return elem;
 }
 
-std::vector<float> Rotate_3Dvector_random(std::vector<float> vect)
+std::vector<double> Rotate_3Dvector_random(std::vector<double> vect)
 {
-    float angle = Random() * 2 * 3.14159;
-    float C = cos(angle);
-    float S = sin(angle);
-    std::vector<float> rotvect;
+    double angle = Random() * 2 * 3.14159;
+    double C = cos(angle);
+    double S = sin(angle);
+    std::vector<double> rotvect;
     rotvect.push_back(pow(C, 2) * vect[0] - S * vect[1] + C * S * vect[2]);
     rotvect.push_back(C * S * vect[0] + C * vect[1] + pow(S, 2) * vect[2]);
     rotvect.push_back(-S * vect[0] + C * vect[2]);
     return (rotvect);
 }
 
-float get_ao(float beta, float k)
+double get_ao(double beta, double k)
 {
     int R = 1;
-    float expo = exp(-2 * beta * k);
-    float ao;
-    float x;
+    double expo = exp(-2 * beta * k);
+    double ao;
+    double x;
     while (R == 1)
     {
         x = expo + Random() * (1 - expo);
@@ -58,7 +58,7 @@ float get_ao(float beta, float k)
     return (ao);
 }
 
-std::vector<float> PauliInv(std::vector<float> a)
+std::vector<double> PauliInv(std::vector<double> a)
 {
     for (int d = 1; d < 4; d++)
     {
@@ -67,20 +67,20 @@ std::vector<float> PauliInv(std::vector<float> a)
     return (a);
 }
 
-std::vector<float> Paulimult(std::vector<float> Ma, std::vector<float> Mb)
+std::vector<double> Paulimult(std::vector<double> Ma, std::vector<double> Mb)
 {
-    std::vector<float> MProd;
-    float Ma0 = Ma[0];
-    float Ma1 = Ma[1];
-    float Ma2 = Ma[2];
-    float Ma3 = Ma[3];
+    std::vector<double> MProd;
+    double Ma0 = Ma[0];
+    double Ma1 = Ma[1];
+    double Ma2 = Ma[2];
+    double Ma3 = Ma[3];
 
-    float Mb0 = Mb[0];
-    float Mb1 = Mb[1];
-    float Mb2 = Mb[2];
-    float Mb3 = Mb[3];
+    double Mb0 = Mb[0];
+    double Mb1 = Mb[1];
+    double Mb2 = Mb[2];
+    double Mb3 = Mb[3];
 
-    // float Mb0, Mb1, Mb2, Mb3 = Mb[0], Mb[1], Mb[2], Mb[3];
+    // double Mb0, Mb1, Mb2, Mb3 = Mb[0], Mb[1], Mb[2], Mb[3];
 
     MProd.push_back(Ma0 * Mb0 - (Ma1 * Mb1 + Ma2 * Mb2 + Ma3 * Mb3));
     MProd.push_back((Ma0 * Mb1 + Ma1 * Mb0) - (Ma2 * Mb3 - Ma3 * Mb2));
@@ -93,16 +93,16 @@ int Mod(int a, int b)
     int ret = a % b;
     return ret >= 0 ? ret : ret + b;
 }
-float PauliDet(std::vector<float> Ulink)
+double PauliDet(std::vector<double> Ulink)
 {
     return (pow(Ulink[0], 2) + pow(Ulink[1], 2) + pow(Ulink[2], 2) + pow(Ulink[3], 2));
 }
-float Trace(std::vector<float> Ulist)
+double Trace(std::vector<double> Ulist)
 {
-    std::vector<float> Prod(Ulist.end() - 4, Ulist.end());
+    std::vector<double> Prod(Ulist.end() - 4, Ulist.end());
     for (int u = 4; u < Ulist.size(); u += 4)
     {
-        std::vector<float> vectu(Ulist.end() - u - 4, Ulist.end() - u);
+        std::vector<double> vectu(Ulist.end() - u - 4, Ulist.end() - u);
         Prod = Paulimult(vectu, Prod);
     }
     return (2 * Prod[0]);
