+++ graph.h
@@ -11,6 +11,6 @@ using json = nlohmann::json;
 #ifndef GRAPH
 #define GRAPH
 // struct Lattice;
-std::vector<std::vector<float>> Lattice_calculation(std::vector<float> N, int Ini_mode, int Imax, float Beta, int graphnumb);
-void Graph1(std::vector<int> Nlist, int Iterations, float Beta);
+std::vector<std::vector<double>> Lattice_calculation(std::vector<double> N, int Ini_mode, int Imax, double Beta, int graphnumb);
+void Graph1(std::vector<int> Nlist, int Iterations, double Beta);
 #endif
