+++ lattice.h
@@ -7,7 +7,7 @@
 
 struct Lattice
 {
-    // a_array(int Lsize) : Lsize(Lsize), data(4, std::vector<float>(pow(Lsize, 4) * 16)) {}
+    // a_array(int Lsize) : Lsize(Lsize), data(4, std::vector<double>(pow(Lsize, 4) * 16)) {}
     int Lattice_length;
     bool ini_cond;
     int Numb_cpu;
@@ -15,28 +15,28 @@ struct Lattice
     int totnumb_elements;
     int Nplaq;
     a_array a;
-    std::vector<float> Avplaq_data;
-    std::vector<float> Wloop_data;
+    std::vector<double> Avplaq_data;
+    std::vector<double> Wloop_data;
     Lattice(){};
     Lattice(int Lat_Length_set, bool ini_cond_set, int Numb_cpu_set) : Lattice_length(Lat_Length_set), ini_cond(ini_cond_set), Numb_cpu(Numb_cpu_set), a(Lat_Length_set), totnumb_orientlinks(4 * pow(Lat_Length_set, 4)), totnumb_elements(4 * totnumb_orientlinks), Nplaq(18 * totnumb_orientlinks)
     {
         a.init(ini_cond);
     }
 
-    std::vector<float> Neighprod(int ind1, int ind2, int ind3, int ind4, int d1, int d2, int revd2)
+    std::vector<double> Neighprod(int ind1, int ind2, int ind3, int ind4, int d1, int d2, int revd2)
     {
-        std::vector<float> Ulinks = Neighbours(ind1, ind2, ind3, ind4, d1, d2, 1, pow(-1, revd2));
-        std::vector<float> U3(Ulinks.end() - 4, Ulinks.end());
-        std::vector<float> U2(Ulinks.end() - 8, Ulinks.end() - 4);
-        std::vector<float> U1(Ulinks.end() - 12, Ulinks.end() - 8);
+        std::vector<double> Ulinks = Neighbours(ind1, ind2, ind3, ind4, d1, d2, 1, pow(-1, revd2));
+        std::vector<double> U3(Ulinks.end() - 4, Ulinks.end());
+        std::vector<double> U2(Ulinks.end() - 8, Ulinks.end() - 4);
+        std::vector<double> U1(Ulinks.end() - 12, Ulinks.end() - 8);
 
-        std::vector<float> Uprod = Paulimult(U1, Paulimult(U2, U3));
+        std::vector<double> Uprod = Paulimult(U1, Paulimult(U2, U3));
         return (Uprod);
     }
-    std::vector<float> Neighbours(int ind1, int ind2, int ind3, int ind4, int d1, int d2, int signd1, int signd2, int Wsize = 1)
+    std::vector<double> Neighbours(int ind1, int ind2, int ind3, int ind4, int d1, int d2, int signd1, int signd2, int Wsize = 1)
     {
-        std::vector<float> Ulist;
-        std::vector<float> a_elem;
+        std::vector<double> Ulist;
+        std::vector<double> a_elem;
         std::vector<int> Xini{ind1, ind2, ind3, ind4};
         int w, x, y, z;
         auto Uargs = [this](std::vector<int> Xijkl, int Ws, int dd, int signdd)
@@ -119,9 +119,9 @@ struct Lattice
         }
         return (Ulist);
     }
-    float site_action(int size, int i1, int i2, int i3, int i4, int direct1)
+    double site_action(int size, int i1, int i2, int i3, int i4, int direct1)
     {
-        float S = 12 * (size == 1);
+        double S = 12 * (size == 1);
         for (int revd1 = 0; revd1 < 2; revd1++)
         {
             for (int direct2 = 0; direct2 < 4; direct2++)
@@ -138,10 +138,10 @@ struct Lattice
         }
         return (S);
     }
-    float Action(int size = 1)
+    double Action(int size = 1)
     {
-        float S = 0;
-        float si;
+        double S = 0;
+        double si;
         for (int i1 = 0; i1 < Lattice_length; i1++)
         {
             for (int i2 = 0; i2 < Lattice_length; i2++)
@@ -186,11 +186,11 @@ struct Lattice
             Wloop_data.push_back(Action(WS) / Nplaq);
         }
     }
-    std::vector<float> New_element(float beta, int i1, int i2, int i3, int i4, int d1)
+    std::vector<double> New_element(double beta, int i1, int i2, int i3, int i4, int d1)
     {
-        std::vector<float> New_a_elem;
-        std::vector<float> USum{0, 0, 0, 0};
-        std::vector<float> neighprod;
+        std::vector<double> New_a_elem;
+        std::vector<double> USum{0, 0, 0, 0};
+        std::vector<double> neighprod;
         for (int d = 0; d < 4; d++)
         {
             if (d == d1)
@@ -206,19 +206,19 @@ struct Lattice
                 }
             }
         }
-        float k = sqrt(PauliDet(USum));
+        double k = sqrt(PauliDet(USum));
         New_a_elem.push_back(get_ao(beta, k));
-        float avv = sqrt(1 - pow(New_a_elem[0], 2)) / sqrt(3);
-        std::vector<float> av = {avv, avv, avv};
-        std::vector<float> a_vect = Rotate_3Dvector_random(av);
+        double avv = sqrt(1 - pow(New_a_elem[0], 2)) / sqrt(3);
+        std::vector<double> av = {avv, avv, avv};
+        std::vector<double> a_vect = Rotate_3Dvector_random(av);
         New_a_elem.push_back(a_vect[0]);
         New_a_elem.push_back(a_vect[1]);
         New_a_elem.push_back(a_vect[2]);
         return (New_a_elem);
     }
-    void Touchheat(float Beta)
+    void Touchheat(double Beta)
     {
-        std::vector<float> elem;
+        std::vector<double> elem;
         int linknumb;
         a_array New_a(Lattice_length);
         New_a.init(0);
@@ -245,7 +245,7 @@ struct Lattice
 
         Average_plaquette();
     }
-    void Heatbath(int Iterations, float Beta)
+    void Heatbath(int Iterations, double Beta)
     {
         for (int i = 0; i < Iterations; i++)
         {
